Add Intuition Engine (MACHINE_IE) platform bring-up and 32-bit VDI path
  This change introduces initial Intuition Engine (IE) machine support across
  build, BIOS, BDOS, utility assembly, and VDI. It also hardens several pointer
  and size calculations for -mshort/toolchain compatibility.

  Build/toolchain:
  - Add `LINUX=1` toolchain profile in `Makefile` using `m68k-linux-gnu-gcc-13`.
  - Add Linux-specific flags to avoid builtin size-mismatch issues and GCC ICE
    (`-fno-builtin`, `-fno-ivopts`, `-fno-tree-slsr`).
  - Allow warning compatibility override via `TOOLCHAIN_WARNFIX`.
  - Auto-select `-m68020` when `MACHINE_IE` is active.
  - Add IE platform sources and `EXTRA_BIOS_SRC` injection support.

  Configuration and machine identity:
  - Add MACHINE_IE defaults in `include/config.h`:
    - non-Atari hardware, 4MB STRAM, no TT/ALT RAM
    - advanced CPU enabled
    - stop instruction enabled
    - boot init info duration disabled (`INITINFO_DURATION=0`)
  - Add machine name mapping in `bios/machine.c` (“Intuition Engine”).
  - Document IE target shim files in new `README.md`.

  BIOS bring-up and interrupt behavior:
  - Integrate IE timer install early in boot (`bios/bios.c`) so 200 Hz timing is
    available during startup.
  - Initialize IE video early in boot and adapt `dumpflg` handling for IE.
  - In `bios/screen.c`, set IE fixed mode characteristics:
    - VBL vector setup
    - fixed framebuffer base/size
    - non-changeable resolution
    - color monitor type
    - 32 planes at 640x480
    - 256-color palette reporting
  - In `bios/vectors.S`:
    - skip hardcopy path for IE
    - return via `rte` for IE VBL handler
    - fix timer-C sieve fast-path register preservation (restore `d0`).
  - In `bios/processor.S`, skip 68030 PMMU probe for IE (68020-class behavior).
  - In `bios/xhdi.c`, reject XHNewCookie handler chaining on IE for now.
  - In `bios/disk.c`, guard hard-disk scanning loop by storage feature macros.
  - In `bios/mouse.c`, add IE-specific Initmous path without IKBD writes.
  - In `bios/ikbd.c`, add IE MMIO keyboard polling and absolute mouse delta
    packetization/remapping logic.

  Console/text output on IE framebuffer:
  - In `bios/conout.c`, add RGBA cell rendering for IE text mode, IE-aware
    blank/scroll paths, and disable planar `neg_cell()` inversion in IE mode.

  BDOS and memory arithmetic safety:
  - Fix pointer arithmetic width/sign issues using explicit `ULONG`/`long` casts:
    - `bdos/fsdir.c`, `bdos/fsopnclo.c`
    - `bdos/iumem.c`, `bdos/osmem.c`, `bdos/umem.c`
    - `bdos/kpgmld.c`, `bdos/proc.c`
    - `bios/biosmem.c`
  - Prevent stack-end miscalculation in `proc_go()` under `-mshort` by computing
    via 32-bit address arithmetic.

  Utility/runtime compatibility:
  - In `util/memmove.S` and `util/memset.S`, handle 16-bit `size_t` ABI by
    zero-extending size arguments when needed.
  - In `util/miscasm.S`, add IE-specific stop-until-interrupt behavior compatible
    with IE interrupt levels and simplify SR restore behavior in stop cases.
  - In `util/cookie.c`, disable NVDI cookie probing on IE during bring-up.

  VDI 32-bit IE rendering path:
  - Treat IE as extended-palette capable and export IE palette storage:
    - `vdi/vdi_defs.h`, `vdi/vdi_col.c`.
  - Add IE palette initialization and VDI `vs_color()` handling mapped to RGBA.
  - Update inquiry behavior in `vdi/vdi_control.c` for `v_planes >= 16`.
  - Add IE address helper in `vdi/vdi_inline.h`.
  - Add IE truecolor paths for:
    - filled/contour operations and pixel readback (`vdi/vdi_fill.c`)
    - line/rectangle software blits (`vdi/vdi_line.c`)
    - mouse sprite display/restore (`vdi/vdi_mouse.c`)
    - text blitting (`vdi/vdi_textblit.c`)
    - raster copy/transform (`vdi/vdi_raster.c`) including:
      - 32-bit standard<->device transform
      - `vro_cpyfm` and `vrt_cpyfm` logic ops in chunky RGBA mode
      - bypass of planar setup for 32-plane IE mode.
  - Expand cursor save area for IE truecolor in `include/lineavars.h` and
    `bios/lineavars.S`.

  Overall result:
  - Adds end-to-end IE platform plumbing and makes core memory/math paths robust
    for 16-bit `size_t`/`-mshort` toolchains, while enabling 32-bit chunky VDI
    rendering and input/timer integration on MACHINE_IE.

+++ bdos/fsdir.c
@@ -1595,7 +1595,7 @@ FCB *scan(DND *dnd, const char *n, WORD att, LONG *posp)
 
     if (*posp == -1)
     {       /*  seek to position of found entry  */
-        ixlseek(fd,fd->o_bytnum - sizeof(FCB));
+        ixlseek(fd,fd->o_bytnum - (long)sizeof(FCB));
         return (FCB *)dnd1;
     }
 
@@ -1671,7 +1671,7 @@ static DND *makdnd(DND *p, FCB *fcb)
     swpw(p1->d_strtcl);
     p1->d_drv = p->d_drv;
     p1->d_dirfil = fd;
-    p1->d_dirpos = fd->o_bytnum - sizeof(FCB);
+    p1->d_dirpos = fd->o_bytnum - (long)sizeof(FCB);
     p1->d_td.time = fcb->f_td.time; /* note: DND time/date are  */
     p1->d_td.date = fcb->f_td.date; /*  actually little-endian! */
     memcpy(p1->d_name, fcb->f_name, FNAMELEN);
